requirements-dev.txt
black>=22.10.0
flake8>=5.0.4
isort>=5.10.1
pytest>=7.1.3
# For tests
torchvision
timm
